Refactor Headers to address forward declaration presumed to be cause of invalid use of type error
