+++ src/DropAnalysis.cpp
@@ -46,11 +46,22 @@ email:  dtarb@usu.edu
 #include <queue>
 #include <iostream>
 #include "commonLib.h"
+#include "tiffIO.h"
 #include "linearpart.h"
 #include "createpart.h"
-#include "tiffIO.h"
 #include "DropAnalysis.h"
 using namespace std;
+
+//returns true iff cell at [nrow][ncol] points to cell at [row][col]
+bool pointsToMe(long col, long row, long ncol, long nrow, tdpartition *dirData){
+	short d;
+	if(!dirData->hasAccess(ncol,nrow) || dirData->isNodata(ncol,nrow)){return false;}
+	d=dirData->getData(ncol,nrow,d);
+	if(nrow+d2[d]==row && ncol+d1[d]==col){
+		return true;
+	}
+	return false;
+}
 
 //does the appropriate updates when a junction is found
 void updateAtJunction(short oOut,long i, long ni,long j, long nj, long nx, long ny, tdpartition *dirData,
