+++ src/streamnet.cpp
@@ -93,6 +93,16 @@ OGRSpatialReferenceH  hSRSraster,hSRSshapefile;
 //	doutmidIdx = DBFAddField(dbf1,"DOUT_MID",FTDouble,16,1);
 //}
 
+//returns true iff cell at [nrow][ncol] points to cell at [row][col]
+bool pointsToMe(long col, long row, long ncol, long nrow, tdpartition *dirData) {
+	short d;
+	if (!dirData->hasAccess(ncol, nrow) || dirData->isNodata(ncol, nrow)) { return false; }
+	d = dirData->getData(ncol, nrow, d);
+	if (nrow + d2[d] == row && ncol + d1[d] == col) {
+		return true;
+	}
+	return false;
+}
 
 void createStreamNetShapefile(char *streamnetsrc,char *streamnetlyr,OGRSpatialReferenceH hSRSraster){
